backend/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
psycopg[binary]==3.2.6
django-environ==0.12.0
django-cors-headers==4.7.0
djangorestframework-simplejwt==5.4.0
drf-spectacular==0.28.0
